Fix CSS + comments from PR + numericSelector
