+++ docgen/layouts/api.pug
@@ -3,7 +3,7 @@ mixin nav(currentPath, nav)
     for navItem in nav
       - var className = navItem.path === currentPath ? 'navItem--active' : 'navItem';
       li(class=className)
-        a(href=`/${navItem.path}`)=navItem.file.title
+        a(href=h.publicPath(navItem.path))=navItem.file.title
         if(navItem.children.length > 0) 
           nav(currentPath, navItem.children)
 
@@ -13,7 +13,7 @@ html
     meta(content='IE=edge', http-equiv='X-UA-Compatible')
     meta(charset='utf-8')
     meta(content='width=device-width,initial-scale=1.0,minimum-scale=1.0,maximum-scale=1.0,user-scalable=no', name='viewport')
-    link(rel='icon', href='assets/favicon.png')
+    link(rel='icon', href=h.publicPath('assets/favicon.png'))
     meta(content='IE=edge,chrome=1', http-equiv='X-UA-Compatible')
     title react-instantsearch | #{title}
     link(rel='stylesheet', href=h.publicPath('assets/css/theme.css'))
