+++ docgen/layouts/example.pug
@@ -4,7 +4,7 @@ html
     meta(content='IE=edge', http-equiv='X-UA-Compatible')
     meta(charset='utf-8')
     meta(content='width=device-width,initial-scale=1.0,minimum-scale=1.0,maximum-scale=1.0,user-scalable=no', name='viewport')
-    link(rel='icon', href='assets/favicon.png')
+    link(rel='icon', href=h.publicPath('assets/favicon.png'))
     meta(content='IE=edge,chrome=1', http-equiv='X-UA-Compatible')
     title react-instantsearch | #{title}
     if stylesheets
