+++ package.json
@@ -34,6 +34,7 @@
     "escape-html": "1.0.3",
     "eslint-config-algolia": "5.5.0",
     "gh-pages": "0.11.0",
+    "google-map-react": "0.19.0",
     "lodash": "4.15.0",
     "marked": "0.3.6",
     "memoizee": "0.4.1",
@@ -44,6 +45,7 @@
     "metalsmith-markdown": "0.2.1",
     "metalsmith-navigation": "0.2.9",
     "ms-webpack": "2.0.0-alpha.1",
+    "pug": "2.0.0-beta6",
     "react": "15.3.1",
     "react-docgen": "2.10.0",
     "react-dom": "15.3.1",
@@ -63,9 +65,5 @@
     "node": "^6.5.0",
     "npm": "^3.10.3"
   },
-  "license": "MIT",
-  "devDependencies": {
-    "google-map-react": "0.19.0",
-    "pug": "2.0.0-beta6"
-  }
+  "license": "MIT"
 }
